# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: HR Management System API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /leaves:
    get:
      operationId: listLeaves
      tags: [leaves]
      summary: List leave requests (admin sees all, employee sees own)
      parameters:
        - name: status
          in: query
          required: false
          schema:
            type: string
            enum: [pending, approved, rejected]
      responses:
        "200":
          description: Leave requests
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/LeaveRequest"
    post:
      operationId: applyLeave
      tags: [leaves]
      summary: Apply for a leave
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateLeaveRequest"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LeaveRequest"
  /leaves/{id}:
    patch:
      operationId: updateLeave
      tags: [leaves]
      summary: Update a pending leave request (owner only)
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateLeaveRequest"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LeaveRequest"
    delete:
      operationId: deleteLeave
      tags: [leaves]
      summary: Cancel a pending leave request (owner only)
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Cancelled
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
  /leaves/{id}/approve:
    post:
      operationId: approveLeave
      tags: [leaves]
      summary: Approve a leave (admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Approved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LeaveRequest"
  /leaves/{id}/reject:
    post:
      operationId: rejectLeave
      tags: [leaves]
      summary: Reject a leave (admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Rejected
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LeaveRequest"
...
components:
  schemas:
...
    SuccessResponse:
      type: object
      required: [success]
      properties:
        success: { type: boolean }
...
    LeaveRequest:
      type: object
      required: [id, employeeId, type, startDate, endDate, days, status, reason, mentionedEmployeeIds, attachments]
      properties:
        id: { type: integer }
        employeeId: { type: integer }
        employeeName: { type: string }
        type: { type: string, enum: [sick, casual, annual] }
        startDate: { type: string, format: date }
        endDate: { type: string, format: date }
        days: { type: integer }
        status: { type: string, enum: [pending, approved, rejected] }
        reason: { type: string }
        attachmentUrl: { type: string, nullable: true }
        attachmentName: { type: string, nullable: true }
        attachments:
          type: array
          items:
            $ref: "#/components/schemas/AttachmentItem"
        mentionedEmployeeIds:
          type: array
          items: { type: integer }
        mentionedEmployees:
          type: array
          items:
            $ref: "#/components/schemas/MentionedEmployee"
        appliedAt: { type: string, description: "ISO datetime" }
    CreateLeaveRequest:
      type: object
      required: [type, startDate, endDate, reason]
      properties:
        type: { type: string, enum: [sick, casual, annual] }
        startDate: { type: string, format: date }
        endDate: { type: string, format: date }
        reason: { type: string }
        attachmentUrl: { type: string, nullable: true }
        attachmentName: { type: string, nullable: true }
        attachments:
          type: array
          items:
            $ref: "#/components/schemas/AttachmentItem"
        mentionedEmployeeIds:
          type: array
          items: { type: integer }
    UpdateLeaveRequest:
      type: object
      properties:
        type: { type: string, enum: [sick, casual, annual] }
        startDate: { type: string, format: date }
        endDate: { type: string, format: date }
        reason: { type: string }
        attachmentUrl: { type: string, nullable: true }
        attachmentName: { type: string, nullable: true }
        attachments:
          type: array
          items:
            $ref: "#/components/schemas/AttachmentItem"
        mentionedEmployeeIds:
          type: array
          items: { type: integer }
    AttachmentItem:
      type: object
      required: [url, name]
      properties:
        url: { type: string }
        name: { type: string }
    MentionedEmployee:
      type: object
      required: [id, name]
      properties:
        id: { type: integer }
        name: { type: string }
...